move fit methods out to own module
