+++ src/ibex_bluesky_core/plans/reflectometry/_det_map_align.py
@@ -17,7 +17,6 @@ from matplotlib.axes import Axes
 from ophyd_async.plan_stubs import ensure_connected
 
 from ibex_bluesky_core.callbacks import ISISCallbacks, LiveFit
-from ibex_bluesky_core.callbacks.fitting.fitting_utils import Gaussian
 from ibex_bluesky_core.callbacks.reflectometry import (
     DetMapAngleScanLiveDispatcher,
     DetMapHeightScanLiveDispatcher,
@@ -31,6 +30,7 @@ from ibex_bluesky_core.devices.simpledae.reducers import (
     PeriodSpecIntegralsReducer,
 )
 from ibex_bluesky_core.devices.simpledae.strategies import Waiter
+from ibex_bluesky_core.fitting import Gaussian
 from ibex_bluesky_core.plan_stubs import call_qt_aware
 
 __all__ = ["DetMapAlignResult", "angle_scan_plan", "height_and_angle_scan_plan"]
